requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
azure-storage-blob==12.23.1
azure-identity==1.19.0
xgboost==2.1.2
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
yfinance==0.2.48
ccxt==4.4.29
python-multipart==0.0.17
websockets==14.1
